Refactored directory BaseComponent -> BaseNode
