+++ Main.tscn
@@ -3,7 +3,7 @@
 [ext_resource type="PackedScene" uid="uid://h2iu1fkuw61u" path="res://Scenes/PlayerController/PlayerController.tscn" id="1_bcck4"]
 [ext_resource type="PackedScene" uid="uid://ubc5cusrcpq7" path="res://Scenes/Static/Couch.tscn" id="3_fnfmp"]
 [ext_resource type="PackedScene" uid="uid://bihv5yjq7muuh" path="res://Scenes/Static/SicilianHouse.tscn" id="3_ie86o"]
-[ext_resource type="PackedScene" uid="uid://dwdhndmeqo37" path="res://Scenes/Enemies/BaseComponent/EnemyBase.tscn" id="4_62xft"]
+[ext_resource type="PackedScene" uid="uid://dwdhndmeqo37" path="res://Scenes/Enemies/BaseNode/EnemyBase.tscn" id="4_62xft"]
 
 [node name="Main" type="Node"]
 
@@ -26,6 +26,5 @@ position = Vector2(165, 173)
 [node name="Couch" parent="GameContainer" instance=ExtResource("3_fnfmp")]
 position = Vector2(144, 242)
 
-[node name="EnemyBase" parent="GameContainer" node_paths=PackedStringArray("PathToFollow") instance=ExtResource("4_62xft")]
+[node name="EnemyBase" parent="GameContainer" instance=ExtResource("4_62xft")]
 position = Vector2(393, 231)
-PathToFollow = NodePath("")
