+++ Scenes/Enemies/BaseNode/EnemyBase.tscn
@@ -1,6 +1,6 @@
 [gd_scene load_steps=6 format=3 uid="uid://dwdhndmeqo37"]
 
-[ext_resource type="Script" path="res://Scenes/Enemies/BaseComponent/EnemyBase.cs" id="1_jqb75"]
+[ext_resource type="Script" path="res://Scenes/Enemies/BaseNode/EnemyBase.cs" id="1_jqb75"]
 [ext_resource type="Texture2D" uid="uid://cwto5tbv5no2g" path="res://icon.svg" id="2_rk1ii"]
 
 [sub_resource type="RectangleShape2D" id="RectangleShape2D_73lpi"]
