--- Scenes/Enemies/TestEnemy.cs
@@ -1,9 +1,0 @@
-using Godot;
-using System;
-using TheLoneLanternProject.Scenes.Enemies.BaseComponent;
-
-namespace TheLoneLanternProject.Scenes.Enemies;
-public partial class TestEnemy : EnemyBase
-{
-    
-}
